% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidy_station_list}
\alias{tidy_station_list}
\title{Train station list}
\format{A list with 160 elements.}
\usage{
tidy_station_list
}
\description{
A dataset containing all train stations in Ireland
}
\keyword{dataset}
